BC-633 - store types refactoring (#1962)
