+++ src/store/env-config.ts
@@ -9,45 +9,8 @@ import { rootStore } from "./index";
 import { $axios } from "../utils/api";
 import ContentModule from "@/store/content";
 import FilePathsModule from "@/store/filePaths";
-
-type Status = "pending" | "completed" | "error" | "";
-
-type BusinessError = {
-	statusCode: string;
-	message: string;
-};
-
-export interface Envs {
-	ADMIN_TABLES_DISPLAY_CONSENT_COLUMN: boolean;
-	DOCUMENT_BASE_DIR: string;
-	FALLBACK_DISABLED?: boolean;
-	FEATURE_ADMIN_TOGGLE_STUDENT_LERNSTORE_VIEW_ENABLED?: boolean;
-	FEATURE_ADMIN_TOGGLE_STUDENT_VISIBILITY_ENABLED?: boolean;
-	FEATURE_ES_COLLECTIONS_ENABLED?: boolean;
-	FEATURE_EXTENSIONS_ENABLED?: boolean;
-	FEATURE_LERNSTORE_ENABLED?: boolean;
-	FEATURE_TEAMS_ENABLED?: boolean;
-	FEATURE_SCHOOL_POLICY_ENABLED?: boolean;
-	FEATURE_VIDEOCONFERENCE_ENABLED?: boolean;
-	FEATURE_MATRIX_MESSENGER_ENABLED?: string;
-	I18N__AVAILABLE_LANGUAGES: string;
-	I18N__DEFAULT_LANGUAGE: string;
-	I18N__DEFAULT_TIMEZONE: string;
-	I18N__FALLBACK_LANGUAGE: string;
-	JWT_SHOW_TIMEOUT_WARNING_SECONDS: number;
-	JWT_TIMEOUT_SECONDS: number;
-	MATRIX_MESSENGER__SCHOOL_SETTINGS_VISIBLE?: boolean;
-	MATRIX_MESSENGER__STUDENT_ROOM_CREATION?: boolean;
-	MATRIX_MESSENGER__SCHOOL_ROOM_ENABLED?: boolean;
-	MATRIX_MESSENGER__DISCOVER_URI: string | undefined;
-	MATRIX_MESSENGER__EMBED_URI: string | undefined;
-	MATRIX_MESSENGER__URI: string | undefined;
-	NOT_AUTHENTICATED_REDIRECT_URL: string;
-	ROCKETCHAT_SERVICE_ENABLED?: boolean;
-	SC_THEME: string;
-	SC_TITLE: string;
-	SC_SHORT_TITLE: string;
-}
+import { BusinessError, Status } from "./types/commons";
+import { Envs } from "./types/env-config";
 
 export const requiredVars = {
 	NOT_AUTHENTICATED_REDIRECT_URL: "/login",
@@ -59,7 +22,8 @@ export const requiredVars = {
 export const configsFromEnvironmentVars = {
 	FEATURE_MATRIX_MESSENGER_ENABLED:
 		process.env.FEATURE_MATRIX_MESSENGER_ENABLED,
-	FEATURE_LERNSTORE_ENABLED: process.env.FEATURE_LERNSTORE_ENABLED?.toLowerCase() == 'true' ,
+	FEATURE_LERNSTORE_ENABLED:
+		process.env.FEATURE_LERNSTORE_ENABLED?.toLowerCase() == "true",
 	MATRIX_MESSENGER__EMBED_URI: process.env.MATRIX_MESSENGER__EMBED_URI,
 	MATRIX_MESSENGER__URI: process.env.MATRIX_MESSENGER__URI,
 	MATRIX_MESSENGER__DISCOVER_URI: process.env.MATRIX_MESSENGER__DISCOVER_URI,
