+++ src/store/rooms.ts
@@ -11,29 +11,7 @@ import {
 	DashboardApiFactory,
 	DashboardApiInterface,
 } from "../serverApi/v3/api";
-
-export type RoomsData = {
-	id: string;
-	title: string;
-	shortTitle: string;
-	displayColor: string;
-	xPosition: number;
-	yPosition: number;
-};
-
-type DroppedObject = {
-	from: {
-		x: number;
-		y: number;
-		groupIndex?: number;
-	};
-	to: {
-		x: number;
-		y: number;
-	};
-	item: object;
-};
-
+import { DroppedObject, RoomsData } from "./types/rooms";
 @Module({
 	name: "rooms",
 	namespaced: true,
