+++ src/store/types/auth.ts
@@ -1,0 +1,29 @@
+export type User = {
+	_id: string;
+	__v: number;
+	firstName: string;
+	lastName: string;
+	email: string;
+	updatedAt: string;
+	birthday: string;
+	createdAt: string;
+	preferences: {};
+	schoolId: string;
+	roles: any[];
+	emailSearchValues: string[];
+	firstNameSearchValues: string[];
+	lastNameSearchValues: string[];
+	consent: {};
+	forcePasswordChange: boolean;
+	language: string;
+	fullName: string;
+	id: string;
+	avatarInitials: string;
+	avatarBackgroundColor: string;
+	age: number;
+	displayName: string;
+	permissions: string[];
+	accountId: string;
+	schoolName: string;
+	externallyManaged: boolean;
+};
