+++ src/store/types/commons.ts
@@ -1,0 +1,6 @@
+export type BusinessError = {
+	statusCode: string;
+	message: string;
+};
+
+export type Status = "pending" | "completed" | "error" | "";
